Split ai video engines route helpers
